+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @gitbook/runtime
 
+## 0.23.1
+
+### Patch Changes
+
+- aab7f96: Improve the typing of OAuth functions.
+- 84ea77c: Fix TS error with latest API spec.
+
+## 0.23.0
+
+### Minor Changes
+
+- d7fa48f: Provide core runtime API to handle and dispatch tasks
+
 ## 0.22.0
 
 ### Minor Changes